Fix count with facets filter (#4751)
Fixes: #4659

Currently count with facets filter is not returning correct result. This is because, while calculating count we use pl.Length() function, which returns length of posting list. This will return correct result if there are no facets filter. But if there are facets filter, we need to filter postings from the
posting list which satisfy facets filter.

This PR fixes above issue.