Add script to auto-update badger (#5449)
Fixes DGRAPH-1345

This PR adds a script that can be used to automate badger updates
in dgraph. The script will be used to automatically update badger
in dgraph. The script is intended to be run on CI (team city) and
automatically create PR for the update.

The script performs the following operations
1. Clone Dgraph
2. Run `go get -v github.com/dgraph-io/badger/v2@master`
3. If no change, exit
4. Create a new commit with the changes to go mod and go.sum
5. Commit and result and push it to a branch `update-badger-$(date)`
6. Create a Pull Request on Dgraph repo with the new branch.

**The newly created pull request has to be merged manually**.
The script needs the following environment variables set
- `RELEASE_BRANCHES` => This are the branches on which badger should
    be updated. For ex: `RELEASE_BRANCHES=("master" "release/v1.2" "release/v20.03")`
- `GH_TOKEN` => Used to push changes to github and create the pull request
- `GH_USERNAME` => Used to push the branch to dgraph repo
- `GIT_EMAIL` => The email id used in the commit (author email id)
- `GIT_NAME` => The name used in the commit (author name)