feat(GraphQL): GraphQL health now reported by /probe/graphql (#5802)
Fixes #GRAPHQL-507.

This PR adds /probe/graphql HTTP endpoint which returns HTTP status 200 OK, if the /graphql endpoint is up and running otherwise it returns HTTP status 503 Service Unavailable. It also returns a JSON body of the following form:

{"status":"init/up/updating schema"}
Status init is returned for 503, while the rest of the two statuses return 200 code.