disable compression for WAL (#6255)
This PR disables compression from WAL in badger. WAL is supposed to be small and fast, Cache/compression makes it a
little slow (we save some disk though but its not worth the slowness). Cache has not been disabled as w directory goes to 
about 30GB as found in 1TB write benchmark for dgraph.