Fix(increment): Fix readTs less than minTs (#6317)
Use the readTs for reads via LocalCache. This would ensure that users don't get a lot of ReadTs less than MinTs errors. This PR also fixes the flaky test in increment_test.go.

This has a side-effect on posting list cache. That cache would need to understand versioning to be effective.

Co-authored-by: Daniel Mai <daniel@dgraph.io>