fix(Alpha): Immediately take a snapshot if we don't have one (#6458)
This PR would cause the Alpha leader to immediately trigger a snapshot creation if it doesn't already have one.

This is useful when you bring a new cluster from bulk loader. If you remove an Alpha instance and add a new Alpha instance, the new follower won't get a snapshot if the leader doesn't have one. This PR avoids that by causing the leader to get a snapshot asap.

Fixes DGRAPH-2445