build(jemalloc): Install Jemalloc if missing via Makefile. (#6463)
This PR adds a jemalloc target to makefile to install jemalloc. The make install command will install jemalloc if it's not already installed.

The make jemalloc target currently fails on docker with command sudo not found. I don't have a solution for it.

The change doesn't work on macOS or Windows. Dgraph builds with jemalloc by default on Linux only.

Co-authored-by: Daniel Mai <daniel@dgraph.io>