docs(Query): Update docs to note that only typed data can be deleted using `S * *`(Docs) (#6657)
* Update delete.md

* fix typo

Updated to fix typos

* Update delete.md per feedback that the type restriction is only for deleting data using a wildcard search

Per Michael Compton, the limitation is not as severe as I thought with the initial draft.

* Address wildcard delete with a sub-section in the topic

* Delete delete.md.sb-e9c19abf-QiH8oo

Deleted a file generated by my text editor. Thank you to Michael for catching this!

* Update delete.md

Update to address remaining feedback on this PR.

* Update delete.md

Changed the reference to "triples" back to the correct terminology, updated related statements.

* Update delete.md

Updated the text of the note on line 47 per feedback from Michael C.

* Update delete.md

typo fix