fix(xidmap): Use skiplist for xidmap (#6765)
This PR changes the underlying map for xidmap from trie to memory-mapped
skip list. This skip list is given a memory mapped buffer of fixed size upfront.
It is expected to not exceed the buffer limit. It also adds a couple of benchmarks
for random reads and writes of skip list.