fix(GraphQL): handle filters for enum properly (#6887)
This PR is a breaking change for the `eq` filter on the List of `enum` Types.

This PR fixes bugs in the filter generated for enum types. For the given schema:-
```
enum Status {
   ACTIVE
   DEACTIVATED
}

type VerificationFilter {
    status: [Status]! @search
}
```

the `hash` filter generated for `status` was:
```
input Status_hash {
   eq: [Status]!
   in: [Status]
}
```
whereas, it should be:
```
input Status_hash {
    eq: Status
    in: [Status]
}
```

This PR also fixes incorrect filter generated for `exact` index in List of enum types. 
Earlier it was being generated like:
```
input Status_exact {
	eq: [Status]
	in: [Status]
	le: [Status]
	lt: [Status]
	ge: [Status]
	gt: [Status]
	between: Status
}
```
but it should be like:
```
input Status_exact {
	eq: Status
	in: [Status]
	le: Status
	lt: Status
	ge: Status
	gt: Status
	between: Status
}
```

Breaking change lies in the generation of `eq` filters. Earlier it was `eq: [Status]!` but now it is `eq: Status`. Both are syntactically fine but `eq: [Status]!` conflicts with the expected usage of `eq` filter.