fix(GraphQL): fix panic when querying custom field in fragment on interface (GRAPHQL-734) (#6909)
Fixes GRAPHQL-734.
This PR fixes a panic when a custom field in a type was queried via an interface using a fragment on that type. The fix is to return the field definition instead of just the field name while constructing required fields for a custom field. So that one can find the field's Dgraph alias while constructing the child query during rewriting.