[BREAKING] Fix(OOM): Don't unmarshal pb.Proposals until we need them (#7059)
In worker/draft.go, in Raft.Ready() when we get committed entries, we were unmarshalling them to pb.Proposals upfront. Instead, push them to applyCh as it is, and only when we need them, we unmarshal them. This should fix an OOM issue we see with pb.Proposals unmarshal taking up a lot of memory.

We achieve this by switching proposal.Key which is a string to a uint64 key, which can be put directly in the raft entry.Data in the first 8 bytes. The rest are used to marshal proposal.

Additional changes:
* Bump up max pending proposal size to 256 MB to increase proposal throughput.

Co-authored-by: Ahsan Barkati <ahsanbarkati@gmail.com>
Co-authored-by: NamanJain8 <jnaman806@gmail.com>
Co-authored-by: Daniel Mai <daniel@dgraph.io>