[Breaking] snapshot_after flag moved inside --raft (#7337)
This PR introduces a new SuperFlag struct, which makes flag handling easier. The setting of the default values for the raft flag is moved within the worker package, so Alpha doesn't need to know about it.

The breaking change here is the move of snapshot_after flag to within raft, considering they are logically together.